Skip rxtree building by .rxtree-version

+++ rxtree/build_for_xcode.sh
@@ -2,8 +2,20 @@
 
 ROOT_DIR=$(cd $(dirname $0); pwd)
 
-echo "Building rxtree..."
+cd "$ROOT_DIR/../../../"
 
-cd $ROOT_DIR
-swift build
-cp .build/debug/rxtree "$ROOT_DIR/../../../"
+if [ ! -f .rxtree-version ]; then
+    echo "0.0" > .rxtree-version
+fi
+
+VERSION=$(cat Podfile.lock | grep RxController\ \( | sed -e 's/  - RxController (//g' | sed -e 's/)://g')
+
+if [ $(cat .rxtree-version) = ${VERSION} ] && [ -f rxtree ]; then
+    echo "rxtree ${VERSION} existed, skip."
+else
+	echo "Building rxtree $VERSION..."
+	echo $VERSION > .rxtree-version
+	cd $ROOT_DIR
+	swift build
+	cp .build/debug/rxtree "$ROOT_DIR/../../../"
+fi
